+++ payment/src/main/java/com/example/payment/infrastructure/client/dto/response/OrderPaymentValidationOrderItemResponse.java
@@ -1,10 +1,11 @@
 package com.example.payment.infrastructure.client.dto.response;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record OrderPaymentValidationOrderItemResponse(
         UUID orderItemId,
         UUID sellerId,
-        Long lineAmount
+        BigDecimal lineAmount
 ) {
 }
